chapter_001/src/main/java/assertjexeption/SimpleModel.java
package assertjexeption;

/**
 * Утверждения с исключениями [#504886]
 * (есть тесты - там см примеры)
 *
 */
public class SimpleModel {
    private String name = "";

    public String getName() {
        if (name.length() == 0) {
            throw new IllegalArgumentException();
        }
        return name;
    }

    public void setName(String word, int number) {
        if (word.length() != number) {
            throw new IllegalArgumentException(
                    String.format("this word: %s has length not equal to : %s", word, number)
            );
        }
        this.name = word;
    }
}